# openapi/artifact-service.openapi.yaml
openapi: 3.0.4
components:
  schemas:
    ArtifactDto:
      type: object
      properties:
        artifactId:
          $ref: '#/components/schemas/UUID'
        size:
          type: integer
          format: int64
        originalName:
          type: string
        contentType:
          type: string
        createdAt:
          $ref: '#/components/schemas/Instant'
        ownerId:
          $ref: '#/components/schemas/UUID'
...
    BaseTaskSummary:
      type: object
      properties:
        total:
          type: integer
          format: int32
        queued:
          type: integer
          format: int32
        running:
          type: integer
          format: int32
        succeeded:
          type: integer
          format: int32
        failed:
          type: integer
          format: int32
        timedOut:
          type: integer
          format: int32
        tasksPerSec:
          type: number
          format: double
    EngineTaskStatus:
      type: string
      enum:
      - STARTING
      - RUNNING
      - SUCCEEDED
      - FAILED
      - TIMED_OUT
...
    ExecutionType:
      type: string
      enum:
      - stateless
      - swarm-sync
...
    Instant:
      type: string
      format: date-time=
...
    LocalDateTime:
      type: string
      format: date-time
...
    MicrotaskRunStatus:
      type: string
      enum:
      - QUEUED
      - STARTING
      - RUNNING
      - SUCCEEDED
      - FAILED
      - TIMED_OUT
...
    OutputList:
      type: object
      properties:
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactDto'
...
    PaginatedResponseTaskRunLight:
      type: object
      properties:
        total:
          type: integer
          format: int32
        pageCount:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        pageIndex:
          type: integer
          format: int32
        records:
          type: array
          items:
            $ref: '#/components/schemas/TaskRunLight'
...
    StatelessMicrotaskRun:
      type: object
      properties:
        taskId:
          $ref: '#/components/schemas/UUID'
        microtaskId:
          $ref: '#/components/schemas/UUID'
        displayIndex:
          type: integer
          format: int32
        status:
          $ref: '#/components/schemas/MicrotaskRunStatus'
        createdAt:
          $ref: '#/components/schemas/LocalDateTime'
        startedAt:
          $ref: '#/components/schemas/LocalDateTime'
        finishedAt:
          $ref: '#/components/schemas/LocalDateTime'
        runDeadline:
          $ref: '#/components/schemas/LocalDateTime'
        runTimeoutSeconds:
          type: integer
          format: int64
        reason:
          type: string
    StatelessMicrotaskState:
      type: object
      properties:
        microtaskId:
          $ref: '#/components/schemas/UUID'
        displayIndex:
          type: integer
          format: int32
        status:
          $ref: '#/components/schemas/MicrotaskRunStatus'
    StatelessTaskState:
      type: object
      properties:
        taskId:
          $ref: '#/components/schemas/UUID'
        seq:
          type: integer
          format: int64
        kind:
          type: string
        status:
          $ref: '#/components/schemas/EngineTaskStatus'
        taskStatus:
          $ref: '#/components/schemas/TaskStatus'
        summary:
          $ref: '#/components/schemas/BaseTaskSummary'
        microtasks:
          type: array
          items:
            $ref: '#/components/schemas/StatelessMicrotaskState'
    SwarmAgent:
      type: object
      properties:
        agentId:
          $ref: '#/components/schemas/UUID'
        taskId:
          $ref: '#/components/schemas/UUID'
        agentIndex:
          type: integer
          format: int32
        inputData:
          type: string
        stateData:
          type: string
        statePhase:
          $ref: '#/components/schemas/SwarmPhase'
        stateIteration:
          type: integer
          format: int32
    SwarmAgentState:
      type: object
      properties:
        agentId:
          $ref: '#/components/schemas/UUID'
        displayIndex:
          type: integer
          format: int32
        status:
          $ref: '#/components/schemas/MicrotaskRunStatus'
        currentIteration:
          type: integer
          format: int32
        currentPhase:
          $ref: '#/components/schemas/SwarmPhase'
    SwarmMicrotaskRun:
      type: object
      properties:
        taskId:
          $ref: '#/components/schemas/UUID'
        microtaskId:
          $ref: '#/components/schemas/UUID'
        displayIndex:
          type: integer
          format: int32
        status:
          $ref: '#/components/schemas/MicrotaskRunStatus'
        createdAt:
          $ref: '#/components/schemas/LocalDateTime'
        startedAt:
          $ref: '#/components/schemas/LocalDateTime'
        finishedAt:
          $ref: '#/components/schemas/LocalDateTime'
        runDeadline:
          $ref: '#/components/schemas/LocalDateTime'
        runTimeoutSeconds:
          type: integer
          format: int64
        reason:
          type: string
        agentId:
          $ref: '#/components/schemas/UUID'
        phase:
          $ref: '#/components/schemas/SwarmPhase'
        iteration:
          type: integer
          format: int32
    SwarmPhase:
      type: string
      enum:
      - INIT
      - STEP
      - FINISH
      - FAILED
...
    SwarmTaskState:
      type: object
      properties:
        taskId:
          $ref: '#/components/schemas/UUID'
        seq:
          type: integer
          format: int64
        kind:
          type: string
        status:
          $ref: '#/components/schemas/EngineTaskStatus'
        taskStatus:
          $ref: '#/components/schemas/TaskStatus'
        summary:
          $ref: '#/components/schemas/SwarmTaskSummary'
        agentStates:
          type: array
          items:
            $ref: '#/components/schemas/SwarmAgentState'
    SwarmTaskSummary:
      type: object
      properties:
        total:
          type: integer
          format: int32
        queued:
          type: integer
          format: int32
        running:
          type: integer
          format: int32
        succeeded:
          type: integer
          format: int32
        failed:
          type: integer
          format: int32
        timedOut:
          type: integer
          format: int32
        tasksPerSec:
          type: number
          format: double
        currentIteration:
          type: integer
          format: int32
        currentPhase:
          $ref: '#/components/schemas/SwarmPhase'
...
    TaskCreateRequest:
      type: object
      properties:
        jarId:
          $ref: '#/components/schemas/UUID'
        inputId:
          $ref: '#/components/schemas/UUID'
        configId:
          $ref: '#/components/schemas/UUID'
    TaskRunDto:
      type: object
      properties:
        taskId:
          $ref: '#/components/schemas/UUID'
        projectId:
          $ref: '#/components/schemas/UUID'
        launchedByUser:
          $ref: '#/components/schemas/UUID'
        executionType:
          $ref: '#/components/schemas/ExecutionType'
        status:
          $ref: '#/components/schemas/TaskStatus'
        jarId:
          $ref: '#/components/schemas/UUID'
        jarAlias:
          type: string
        inputId:
          $ref: '#/components/schemas/UUID'
        inputAlias:
          type: string
        configId:
          $ref: '#/components/schemas/UUID'
        configAlias:
          type: string
        createdAt:
          $ref: '#/components/schemas/Instant'
        startedAt:
          $ref: '#/components/schemas/Instant'
        finishedAt:
          $ref: '#/components/schemas/Instant'
        doneAt:
          $ref: '#/components/schemas/Instant'
    TaskRunLight:
      type: object
      properties:
        taskId:
          $ref: '#/components/schemas/UUID'
        projectId:
          $ref: '#/components/schemas/UUID'
        launchedByUser:
          $ref: '#/components/schemas/UUID'
        status:
          $ref: '#/components/schemas/TaskStatus'
        createdAt:
          $ref: '#/components/schemas/Instant'
        doneAt:
          $ref: '#/components/schemas/Instant'
    TaskStatus:
      type: string
      enum:
      - CREATED
      - STARTING
      - RUNNING
      - FINISHED
      - TIME_OUT
      - CANCELED
      - FAILED
      - DONE
...
    UUID:
      type: string
      format: uuid
      pattern: '[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access JWT, передаваемый в заголовке Authorization: Bearer <token>.'
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refreshJWT
      description: Refresh JWT, передаваемый в cookie refreshJWT.
paths:
...
  /artifact-service/api/projects/{projectId}/tasks:
    get:
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponseTaskRunLight'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
      summary: Получение списка задач
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Возвращает страницу task run проекта.
      operationId: artifact_getTasks
    post:
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
        required: true
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRunDto'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
        '400':
          description: Некорректный запрос или ошибка валидации
      summary: Создание запуска задачи
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Создаёт task run на основе jar, input и execution config.
      operationId: artifact_createTask
    servers:
    - url: http://localhost:8082
  /artifact-service/api/projects/{projectId}/tasks/{taskId}:
    get:
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRunDto'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
      summary: Получение задачи
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Возвращает сведения о task run.
      operationId: artifact_getTask
    patch:
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Операция выполнена успешно
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
      summary: Отмена задачи
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Отправляет запрос на отмену task run.
      operationId: artifact_cancelTask
    servers:
    - url: http://localhost:8082
  /artifact-service/api/projects/{projectId}/tasks/{taskId}/agents/{agentId}:
    get:
      parameters:
      - name: agentId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwarmAgent'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
      summary: Получение swarm-агента
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Возвращает сведения об агенте swarm-задачи.
      operationId: artifact_getSwarmAgent
    servers:
    - url: http://localhost:8082
  /artifact-service/api/projects/{projectId}/tasks/{taskId}/microtasks/stateless/{microtaskId}:
    get:
      parameters:
      - name: microtaskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatelessMicrotaskRun'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
      summary: Получение stateless-микрозадачи
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Возвращает сведения о stateless-микрозадаче.
      operationId: artifact_getStatelessMicrotask
    servers:
    - url: http://localhost:8082
  /artifact-service/api/projects/{projectId}/tasks/{taskId}/microtasks/swarm/{microtaskId}:
    get:
      parameters:
      - name: microtaskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwarmMicrotaskRun'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
      summary: Получение swarm-микрозадачи
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Возвращает сведения о swarm-микрозадаче.
      operationId: artifact_getSwarmMicrotask
    servers:
    - url: http://localhost:8082
  /artifact-service/api/projects/{projectId}/tasks/{taskId}/outputs:
    get:
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputList'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
      summary: Получение output-файлов задачи
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Возвращает список выходных файлов задачи.
      operationId: artifact_getTaskOutputs
    servers:
    - url: http://localhost:8082
  /artifact-service/api/projects/{projectId}/tasks/{taskId}/stats/stateless:
    get:
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatelessTaskState'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
      summary: Получение stateless-состояния
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Возвращает агрегированное состояние stateless-задачи.
      operationId: artifact_getStatelessTaskState
    servers:
    - url: http://localhost:8082
  /artifact-service/api/projects/{projectId}/tasks/{taskId}/stats/swarm:
    get:
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: taskId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Операция выполнена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwarmTaskState'
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
      summary: Получение swarm-состояния
      tags:
      - Задачи
      security:
      - BearerAuth: []
      description: Возвращает агрегированное состояние swarm-задачи.
      operationId: artifact_getSwarmTaskState
    servers:
    - url: http://localhost:8082
info:
  title: MinionFlow Artifact Service API
  version: 1.0-SNAPSHOT
  description: 'API сервиса artifact-service: управление jar-артефактами, входными файлами, execution config, task run, логами
    и output-файлами.'
servers:
- url: http://localhost:8082
  description: Локальный запуск artifact-service
...